Make the schema lane actually run, and fix what it found
The lane has never reached pgTAP. On main it dies at
`20260820161957_schedule_synthesis_worker.sql` with
"vault secrets synthesis_worker_key/synthesis_worker_url are missing"
(run 34032987810), so `supabase test db` was never invoked. The
handoff worried that seven `like()` assertions might never execute.
It was 274 assertions, in fifteen files, none of which had ever run
in CI.

With the hosted-readiness gate in place the migrations complete and
pg_prove runs for the first time. What it found:

`authenticated` holds INSERT on `relationship_events`, `seasons` and
`contextual_suggestions`. Those tables were given `grant select` and
nothing else, which reads as read-only, but Supabase's bootstrap
grants all privileges on new public tables by default and no migration
took the writes back. Nothing is forgeable today — RLS is on and each
table carries a select-only policy — but one policy is the only thing
standing between a client and a fabricated Season. `public.plans`
already set the precedent; the new migration follows it. Every write
comes from a `security definer` RPC and is unaffected.

Four files aborted partway and reported the assertions before the
abort as passing, so the remainder ran silently to nothing:

- `native_product.test.sql:77` and `field_couple_contract.test.sql:175`
  compare a smallint to an integer and a `name` to `text`. pgTAP has
  no such `is()`, the file stops there, and in native_product that is
  what kept `:347` — one of the converted assertions — unreached.
- Three files build a temp fixture as the owner and read it back as
  `authenticated`, which has no privilege on it.
- `private_answers_shared_directions.test.sql` seeds a plan by direct
  insert while holding `authenticated`, which the product forbids and
  `native_product.test.sql:52` asserts it forbids. Seeded as owner now.
- `yours_personal_space.test.sql` adds a second couple member without
  `member_slot`, which is NOT NULL and is the A/B contract.

`supabase test db` runs pg_prove over everything under `supabase/tests`,
so two local helpers and two probes were being run as tests and failing
the lane. They move to `supabase/probes/`.

+++ supabase/tests/native_product.test.sql
@@ -68,7 +68,9 @@ select ialike(
 );
 select is(
   (
-    select pronargdefaults
+    -- `pronargdefaults` is smallint; the literal below is integer, and
+    -- pgTAP's `is()` needs both sides to agree or the file aborts here.
+    select pronargdefaults::integer
     from pg_proc
     where oid = 'public.submit_response(uuid,text,boolean,text)'::regprocedure
   ),
@@ -77,7 +79,9 @@ select is(
 );
 select is(
   (
-    select pronargdefaults
+    -- `pronargdefaults` is smallint; the literal below is integer, and
+    -- pgTAP's `is()` needs both sides to agree or the file aborts here.
+    select pronargdefaults::integer
     from pg_proc
     where oid = 'public.resolve_insight(uuid,text,text)'::regprocedure
   ),
